Add tests & move perfomance tests in separate file
